Update to use api as source of truth
