+++ src/applications/vaos/appointment-list/pages/RequestedAppointmentsPage/RequestedAppointmentsPage.unit.spec.js
@@ -2,6 +2,7 @@ import { mockFetch } from '@department-of-veterans-affairs/platform-testing/help
 import { expect } from 'chai';
 import MockDate from 'mockdate';
 import moment from 'moment';
+import { within } from '@testing-library/dom';
 import React from 'react';
 import reducers from '../../../redux/reducer';
 import { mockVAOSAppointmentsFetch } from '../../../tests/mocks/helpers';
@@ -16,10 +17,11 @@ const initialStateVAOSService = {
   featureToggles: {
     vaOnlineSchedulingCancel: true,
     vaOnlineSchedulingVAOSServiceRequests: true,
+    vaOnlineSchedulingFeSourceOfTruth: true,
   },
 };
 
-describe('VAOS Component: RequestedAppointmentsList', () => {
+describe('VAOS Component: RequestedAppointmentsPage', () => {
   beforeEach(() => {
     mockFetch();
     MockDate.set(getTestDate());
@@ -29,6 +31,145 @@ describe('VAOS Component: RequestedAppointmentsList', () => {
     MockDate.reset();
   });
 
+  it('should show va request', async () => {
+    // Given a veteran has VA appointment request
+    const startDate = moment.utc();
+    const appointment = getVAOSRequestMock();
+    appointment.id = '1234';
+    appointment.attributes = {
+      comment: 'A message from the patient',
+      contact: {
+        telecom: [
+          { type: 'phone', value: '2125551212' },
+          { type: 'email', value: 'veteranemailtest@va.gov' },
+        ],
+      },
+      kind: 'clinic',
+      locationId: '983',
+      location: {
+        id: '983',
+        type: 'appointments',
+        attributes: {
+          id: '983',
+          vistaSite: '983',
+          name: 'Cheyenne VA Medical Center',
+          lat: 39.744507,
+          long: -104.830956,
+          phone: { main: '307-778-7550' },
+        },
+      },
+      id: '1234',
+      preferredTimesForPhoneCall: ['Morning'],
+      reasonCode: {
+        coding: [{ code: 'Routine Follow-up' }],
+        text: 'A message from the patient',
+      },
+      requestedPeriods: [
+        {
+          start: moment(startDate)
+            .add(3, 'days')
+            .format('YYYY-MM-DDTHH:mm:ss[Z]'),
+        },
+        {
+          start: moment(startDate)
+            .add(4, 'days')
+            .format('YYYY-MM-DDTHH:mm:ss[Z]'),
+        },
+      ],
+      serviceType: '323',
+      start: null,
+      status: 'proposed',
+      pending: true,
+    };
+    // And developer is using the v2 API
+    mockVAOSAppointmentsFetch({
+      start: moment()
+        .subtract(120, 'days')
+        .format('YYYY-MM-DD'),
+      end: moment()
+        .add(2, 'days')
+        .format('YYYY-MM-DD'),
+      statuses: ['proposed', 'cancelled'],
+      requests: [appointment],
+    });
+
+    // When veteran selects the Requested dropdown selection
+    const screen = renderWithStoreAndRouter(<RequestedAppointmentsPage />, {
+      initialState: initialStateVAOSService,
+      reducers,
+    });
+    // Then it should display the requested appointments
+    expect(await screen.findByText('Primary care')).to.be.ok;
+    expect(await screen.findByText('Cheyenne VA Medical Center')).to.be.ok;
+    expect(screen.queryByText(/You don’t have any appointments/i)).not.to.exist;
+    expect(screen.baseElement).to.contain.text(
+      'Appointments that you request will show here until staff review and schedule them.',
+    );
+  });
+  it('should show cc request', async () => {
+    // Given a veteran has CC appointment request
+    // practitioners.id is same as practitioners.identifier
+    const startDate = moment.utc();
+    const ccAppointmentRequest = getVAOSRequestMock();
+    ccAppointmentRequest.id = '1234';
+    ccAppointmentRequest.attributes = {
+      comment: 'A message from the patient',
+      contact: {
+        telecom: [
+          { type: 'phone', value: '2125551212' },
+          { type: 'email', value: 'veteranemailtest@va.gov' },
+        ],
+      },
+      kind: 'cc',
+      locationId: '983GC',
+      id: '1234',
+      practitioners: [{ id: [{ value: '123' }] }],
+      preferredTimesForPhoneCall: ['Morning'],
+      reasonCode: {
+        coding: [{ code: 'Routine Follow-up' }],
+        text: 'A message from the patient',
+      },
+      requestedPeriods: [
+        {
+          start: moment(startDate)
+            .add(3, 'days')
+            .format('YYYY-MM-DDTHH:mm:ss[Z]'),
+        },
+        {
+          start: moment(startDate)
+            .add(4, 'days')
+            .format('YYYY-MM-DDTHH:mm:ss[Z]'),
+        },
+      ],
+      serviceType: '203',
+      status: 'proposed',
+      pending: true,
+    };
+
+    // And developer is using the v2 API
+    mockVAOSAppointmentsFetch({
+      start: moment()
+        .subtract(120, 'days')
+        .format('YYYY-MM-DD'),
+      end: moment()
+        .add(2, 'days')
+        .format('YYYY-MM-DD'),
+      statuses: ['proposed', 'cancelled'],
+      requests: [ccAppointmentRequest],
+    });
+    // When veteran selects the Requested dropdown selection
+    const screen = renderWithStoreAndRouter(<RequestedAppointmentsPage />, {
+      initialState: initialStateVAOSService,
+      reducers,
+    });
+
+    expect(await screen.findByText('Audiology and speech')).to.be.ok;
+    expect(screen.baseElement).to.contain.text('Community care');
+    expect(screen.queryByText(/You don’t have any appointments/i)).not.to.exist;
+    expect(screen.baseElement).to.contain.text(
+      'Appointments that you request will show here until staff review and schedule them.',
+    );
+  });
   it('should display pending and canceled appointments grouped', async () => {
     // And a veteran has VA appointment request
     const startDate = moment.utc();
@@ -84,6 +225,7 @@ describe('VAOS Component: RequestedAppointmentsList', () => {
       serviceType: '323',
       start: null,
       status: 'proposed',
+      pending: true,
     };
     const canceledAppointment = {
       ...appointment,
@@ -117,11 +259,118 @@ describe('VAOS Component: RequestedAppointmentsList', () => {
     // Then it should display the requested appointments
     expect(await screen.findByText('Primary care')).to.be.ok;
 
-    // And it hsould display the cancelled appointments
+    // And it should display the cancelled appointments
     expect(screen.getByRole('heading', { level: 2, name: 'Canceled requests' }))
       .to.be.ok;
     expect(screen.getByText('These appointment requests have been canceled.'))
       .to.be.ok;
+  });
+  it('should display request sorted by create date in descending order', async () => {
+    // Given a veteran has VA appointment request
+    const startDate = moment.utc();
+    const appointment = getVAOSRequestMock();
+    appointment.id = '1234';
+    appointment.attributes = {
+      comment: 'A message from the patient',
+      contact: {
+        telecom: [
+          { type: 'phone', value: '2125551212' },
+          { type: 'email', value: 'veteranemailtest@va.gov' },
+        ],
+      },
+      kind: 'clinic',
+      locationId: '983',
+      location: {
+        id: '983',
+        type: 'appointments',
+        attributes: {
+          id: '983',
+          vistaSite: '983',
+          name: 'Cheyenne VA Medical Center',
+          lat: 39.744507,
+          long: -104.830956,
+          phone: { main: '307-778-7550' },
+          physicalAddress: {
+            line: ['2360 East Pershing Boulevard'],
+            city: 'Cheyenne',
+            state: 'WY',
+            postalCode: '82001-5356',
+          },
+        },
+      },
+
+      id: '1234',
+      preferredTimesForPhoneCall: ['Morning'],
+      reasonCode: {
+        coding: [{ code: 'Routine Follow-up' }],
+        text: 'A message from the patient',
+      },
+      requestedPeriods: [
+        {
+          start: moment(startDate)
+            .add(3, 'days')
+            .format('YYYY-MM-DDTHH:mm:ss[Z]'),
+        },
+        {
+          start: moment(startDate)
+            .add(4, 'days')
+            .format('YYYY-MM-DDTHH:mm:ss[Z]'),
+        },
+      ],
+      serviceType: 'primaryCare',
+      start: null,
+      status: 'proposed',
+      created: moment()
+        .subtract(2, 'months')
+        .format('YYYY-MM-DDTHH:mm:ss'),
+      pending: true,
+    };
+    const appointment2 = {
+      ...appointment,
+      attributes: {
+        ...appointment.attributes,
+        created: moment()
+          .clone()
+          .subtract(1, 'month')
+          .format('YYYY-MM-DDTHH:mm:ss'),
+        serviceType: 'audiology',
+      },
+    };
+    const appointment3 = {
+      ...appointment,
+      attributes: {
+        ...appointment.attributes,
+        created: moment().format('YYYY-MM-DDTHH:mm:ss'),
+        serviceType: 'optometry',
+      },
+    };
+
+    // And developer is using the v2 API
+    mockVAOSAppointmentsFetch({
+      start: moment()
+        .subtract(120, 'days')
+        .format('YYYY-MM-DD'),
+      end: moment()
+        .add(2, 'days')
+        .format('YYYY-MM-DD'),
+      statuses: ['proposed', 'cancelled'],
+      requests: [appointment, appointment2, appointment3],
+    });
+
+    // When veteran selects the Requested dropdown selection
+    const screen = renderWithStoreAndRouter(<RequestedAppointmentsPage />, {
+      initialState: initialStateVAOSService,
+      reducers,
+    });
+
+    // Then it should display the requested appointments sorted by create date in decending order.
+    expect(await screen.findByText('Primary care')).to.be.ok;
+
+    const links = screen.getAllByRole('listitem');
+    expect(links.length).to.equal(3);
+    expect(within(links[0]).getByText('Optometry')).to.be.ok;
+    expect(within(links[1]).getByText('Audiology and speech')).to.be.ok;
+    expect(within(links[2]).getByText('Primary care')).to.be.ok;
   });
 
   it('should display pending appointments when there are no canceled appointments', async () => {
@@ -179,6 +428,7 @@ describe('VAOS Component: RequestedAppointmentsList', () => {
       serviceType: '323',
       start: null,
       status: 'proposed',
+      pending: true,
     };
 
     // And developer is using the v2 API
@@ -220,7 +470,7 @@ describe('VAOS Component: RequestedAppointmentsList', () => {
     ).not.to.be.ok;
   });
 
-  it('should dispaly no appointments alert when there are no pending or cancelled appointments', async () => {
+  it('should display no appointments alert when there are no pending or cancelled appointments', async () => {
     // And a veteran has no pending or canceled appointment request
     // And developer is using the v2 API
     mockVAOSAppointmentsFetch({
@@ -306,6 +556,7 @@ describe('VAOS Component: RequestedAppointmentsList', () => {
       serviceType: '323',
       start: null,
       status: 'cancelled',
+      pending: true,
     };
 
     // And developer is using the v2 API
@@ -345,5 +596,104 @@ describe('VAOS Component: RequestedAppointmentsList', () => {
         name: /You don’t have any/,
       }),
     ).to.be.ok;
+  });
+  it('should show error message when request fails', async () => {
+    mockVAOSAppointmentsFetch({ error: true });
+
+    const screen = renderWithStoreAndRouter(<RequestedAppointmentsPage />, {
+      initialState: initialStateVAOSService,
+      reducers,
+    });
+
+    expect(
+      await screen.findByText(
+        /We’re having trouble getting your appointment requests/i,
+      ),
+    ).to.be.ok;
+  });
+
+  describe('When FE Source of Truth is off', () => {
+    const defaultState = {
+      featureToggles: {
+        ...initialStateVAOSService.featureToggles,
+        vaOnlineSchedulingFeSourceOfTruth: false,
+      },
+    };
+
+    it('should show va request', async () => {
+      // Given a veteran has VA appointment request
+      const startDate = moment.utc();
+      const appointment = getVAOSRequestMock();
+      appointment.id = '1234';
+      appointment.attributes = {
+        comment: 'A message from the patient',
+        contact: {
+          telecom: [
+            { type: 'phone', value: '2125551212' },
+            { type: 'email', value: 'veteranemailtest@va.gov' },
+          ],
+        },
+        kind: 'clinic',
+        locationId: '983',
+        location: {
+          id: '983',
+          type: 'appointments',
+          attributes: {
+            id: '983',
+            vistaSite: '983',
+            name: 'Cheyenne VA Medical Center',
+            lat: 39.744507,
+            long: -104.830956,
+            phone: { main: '307-778-7550' },
+          },
+        },
+        id: '1234',
+        preferredTimesForPhoneCall: ['Morning'],
+        reasonCode: {
+          coding: [{ code: 'Routine Follow-up' }],
+          text: 'A message from the patient',
+        },
+        requestedPeriods: [
+          {
+            start: moment(startDate)
+              .add(3, 'days')
+              .format('YYYY-MM-DDTHH:mm:ss[Z]'),
+          },
+          {
+            start: moment(startDate)
+              .add(4, 'days')
+              .format('YYYY-MM-DDTHH:mm:ss[Z]'),
+          },
+        ],
+        serviceType: '323',
+        start: null,
+        status: 'proposed',
+      };
+      // And developer is using the v2 API
+      mockVAOSAppointmentsFetch({
+        start: moment()
+          .subtract(120, 'days')
+          .format('YYYY-MM-DD'),
+        end: moment()
+          .add(2, 'days')
+          .format('YYYY-MM-DD'),
+        statuses: ['proposed', 'cancelled'],
+        requests: [appointment],
+      });
+
+      // When veteran selects the Requested dropdown selection
+      const screen = renderWithStoreAndRouter(<RequestedAppointmentsPage />, {
+        initialState: defaultState,
+        reducers,
+      });
+      // Then it should display the requested appointments
+      expect(await screen.findByText('Primary care')).to.be.ok;
+      expect(await screen.findByText('Cheyenne VA Medical Center')).to.be.ok;
+      expect(screen.queryByText(/You don’t have any appointments/i)).not.to
+        .exist;
+      expect(screen.baseElement).to.contain.text(
+        'Appointments that you request will show here until staff review and schedule them.',
+      );
+    });
   });
 });
