+++ src/applications/vaos/services/appointment/index.unit.spec.js
@@ -551,7 +551,6 @@ describe('VAOS Services: Appointment ', () => {
         startDate,
         endDate,
         useV2: true,
-        useFeSourceOfTruth: true,
       });
 
       // Then expect a CC appointment request to be returned.
