+++ src/applications/vaos/services/appointment/transformers.js
@@ -114,9 +114,10 @@ export function transformVAOSAppointment(appt, useFeSourceOfTruth) {
   const isVideo = appt.kind === 'telehealth' && !!appt.telehealth?.vvsKind;
   const isAtlas = !!appt.telehealth?.atlas;
   const isPast = isPastAppointment(appt);
-  const isRequest =
-    appointmentType === APPOINTMENT_TYPES.request ||
-    appointmentType === APPOINTMENT_TYPES.ccRequest;
+  const isRequest = useFeSourceOfTruth
+    ? appt.pending
+    : appointmentType === APPOINTMENT_TYPES.request ||
+      appointmentType === APPOINTMENT_TYPES.ccRequest;
   const isUpcoming = useFeSourceOfTruth
     ? appt.future
     : isFutureAppointment(appt, isRequest);
